feat: add update functionality
It's not pretty, still needs to be pretty, but it works. I need to
figure out the styles....it'll take time. Probably also needs a refactor
because there' duplicate code between the add and update commands

I moved the 'q' keybind to the list, for the same reason as the 'a'
keybind.